Moving packages around to follow conventions
