+++ src/main/java/w16cs350/controller/cli/parser/patterns/structural/UncouplePatternMatcher.java
@@ -1,11 +1,9 @@
-package w16cs350.controller.cli.parser.structural;
+package w16cs350.controller.cli.parser.patterns.structural;
 
-import w16cs350.controller.cli.TrackLocator;
-import w16cs350.controller.cli.parser.PrimitiveDeserializer;
 import w16cs350.controller.cli.parser.patterns.A_IteratingPatternMatcher;
 import w16cs350.controller.cli.parser.patterns.A_PatternMatcher;
 import w16cs350.controller.command.A_Command;
-import w16cs350.controller.command.structural.CommandStructuralLocate;
+import w16cs350.controller.command.structural.CommandStructuralUncouple;
 import w16cs350.support.Assert;
 
 import java.util.ListIterator;
@@ -13,8 +11,8 @@ import java.util.ListIterator;
 /**
  * Created by michael on 3/3/16.
  */
-public class LocatePatternMatcher extends A_IteratingPatternMatcher {
-    public LocatePatternMatcher(A_PatternMatcher parent) {
+public class UncouplePatternMatcher extends A_IteratingPatternMatcher {
+    public UncouplePatternMatcher(A_PatternMatcher parent) {
         super(parent);
     }
 
@@ -22,7 +20,7 @@ public class LocatePatternMatcher extends A_IteratingPatternMatcher {
     protected boolean isMatch(ListIterator<String> tok) {
         String token = tok.next();
         tok.previous();
-        return token.equals("LOCATE");
+        return token.equals("UNCOUPLE");
     }
 
     @Override
@@ -32,11 +30,13 @@ public class LocatePatternMatcher extends A_IteratingPatternMatcher {
 
     @Override
     protected A_Command parseCommand(ListIterator<String> tokens) {
-        String stock = tokens.next();
-        Assert.isTrue(stock.equals("STOCK"), "Incorrect input, expected: STOCK");
-        String idToken = tokens.next();
-        TrackLocator tl = PrimitiveDeserializer.parseTrackLocator(tokens);
-        return new CommandStructuralLocate(idToken, tl);
+        String stockToken = tokens.next();
+        Assert.isTrue(stockToken.equals("STOCK"), "Incorrect input, expected: STOCK");
+        String id1 = tokens.next();
+        String andToken = tokens.next();
+        Assert.isTrue(andToken.equals("AND"), "Incorrect input, expected: AND");
+        String id2 = tokens.next();
+        return new CommandStructuralUncouple(id1, id2);
     }
 
     @Override
