+++ src/test/java/w16cs350/controller/cli/parser/patterns/structural/LocatePatternMatcherTest.java
@@ -1,0 +1,15 @@
+package w16cs350.controller.cli.parser.patterns.structural;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Created by RowleyJohn on 3/4/2016.
+ */
+public class LocatePatternMatcherTest {
+    @Test
+    public void testLocatePAtternMatcher(){
+
+    }
+}
